requirements.txt
matplotlib==3.1.3
graphviz==0.13.2
tqdm==4.36.1
seaborn==0.9.0
treelib==1.5.5
networkx==2.3
lxml==4.4.1
numpy==1.17.2
scipy==1.3.1
chardet==3.0.4
pandas==0.25.1
Pillow==8.4.0
